requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
idna==3.4
kaleido==0.2.1
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
sdmx1==2.7.0
six==1.16.0
tenacity==8.1.0
typing_extensions==4.4.0
urllib3==1.26.14
